pyproject.toml
[project]
name = "ProjectName"
version = "0.1.0"
description = ""
authors = [{ name = "Aluci" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pre-commit (>=4.2.0,<5.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "torchaudio (>=2.7.1,<3.0.0)",
    "dacite (>=1.9.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "notebook (>=7.4.4,<8.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "fire (>=0.7.0,<0.8.0)",
    "ruamel-yaml (>=0.18.15,<0.19.0)",
    "transformers (>=4.57.0,<5.0.0)",
    "accelerate (>=1.10.1)",
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
